refactor admins and modules (#204)
