+++ src/vocgui/admins/__init__.py
@@ -1,0 +1,9 @@
+from .discipline_admin import DisciplineAdmin
+from .training_set_admin import TrainingSetAdmin
+from .document_admin import DocumentAdmin
+
+__all__ = [
+    "DisciplineAdmin",
+    "TrainingSetAdmin",
+    "DocumentAdmin"
+]
