--- src/vocgui/static/archive/manifest.json
@@ -1,22 +1,0 @@
-{
-    "short_name": "ViVo",
-    "name": "Tuer an Tuer - Visual Vocabulary Trainer",
-    "description": "Simply learn vocabularies - no account required",
-    "icons": [
-        {
-            "src": "/static/favicon-192.png",
-            "sizes": "192x192",
-            "type": "image/png"
-        },
-        {
-            "src": "/static/favicon-512.png",
-            "sizes": "512x512",
-            "type": "image/png"
-        }
-    ],
-    "start_url": "",
-    "display": "standalone",
-    "orientation": "portrait",
-    "theme_color": "#ffffff",
-    "background_color": "#ffffff"
-}
