--- src/vocgui/templates/index.html
@@ -1,48 +1,0 @@
-<!DOCTYPE html>
-<html lang="de">
-  <head>
-    <meta charset="utf-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-    <meta name="theme-color" content="#ffffff">
-    <title>Digitalfabrik - Visueller Fachwortschatz-Trainer</title>
-    <link href="/static/bootstrap/css/bootstrap.min.css" rel="stylesheet">
-    <link href="/static/style.css" rel="stylesheet">
-    <link rel="manifest" href="static/manifest.json">
-    <link rel="apple-touch-icon" href="/static/icon.png">
-    <link rel="shortcut icon" type="image/png" href="static/favicon.png">
-  </head>
-  <body class="d-flex flex-column h-100 min-vh-100">
-    <header>
-      <nav class="navbar navbar-expand-md navbar-dark fixed-top bg-dark">
-        <a class="navbar-brand" href="#">Visueller Fachwortschatz-Trainer</a>
-      </nav>
-    </header>
-    <main role="main" class="flex-shrink-0">
-      <div class="container">
-        <div class="row no-gutters">
-          <div class="col-xs-12" style="height:100px;"></div>
-          <div class="col col-lg-5 offset-lg-3 text-center">
-            <h1>Lektion</h1>
-            <div id="div_training_set text-center">
-              <select id="select_training_set" class="form-control center"><option value="0">Bitte wählen</option></select>
-            </div>
-            <div class="col-xs-12" style="height:50px;"></div>
-            <div id="div_ask_document" class="text-center">
-          
-            </div>
-            <div class="col-xs-12" style="height:50px;"></div>
-          </div>
-        </div>
-      </div>
-    </main>
-    <footer class="footer mt-auto py-3">
-      <div class="container">
-        <span class="text-muted"><a href="https://tuerantuer.de/digitalfabrik">Tür an Tür Digitalfabrik gGmbH</a></span>
-      </div>
-    </footer>
-    <script src="/static/jquery.min.js"></script>
-    <script src="/static/bootstrap/js/bootstrap.min.js"></script>
-    <script src="/static/scripts.js"></script>
-  </body>
-</html>
