--- src/vocgui/templates/report.html
@@ -1,231 +1,0 @@
-
-
-<!-- doctype html -->
-<html>
-<head>
-    <meta charset="utf-8">
-    <title>Zusammenfassung - Vokabeltest</title>
-    
-    <style>
-    .report-box {
-        max-width: 800px;
-        margin: auto;
-        padding: 30px;
-        border: 1px solid #eee;
-        box-shadow: 0 0 10px rgba(0, 0, 0, .15);
-        font-size: 16px;
-        line-height: 24px;
-        font-family: 'Helvetica Neue', 'Helvetica', Helvetica, Arial, sans-serif;
-        color: #555;
-    }
-    
-    .report-box table {
-        width: 100%;
-        line-height: inherit;
-        text-align: left;
-    }
-    
-    .report-box table td {
-        padding: 5px;
-        vertical-align: top;
-    }
-    
-    .report-box table tr td:nth-child(2) {
-        text-align: right;
-    }
-    
-    .report-box table tr.top table td {
-        padding-bottom: 20px;
-    }
-    
-    .report-box table tr.top table td.title {
-        font-size: 45px;
-        line-height: 45px;
-        color: #333;
-    }
-    
-    .report-box table tr.information table td {
-        padding-bottom: 40px;
-    }
-    
-    .report-box table tr.heading td {
-        background: #eee;
-        border-bottom: 1px solid #ddd;
-        font-weight: bold;
-    }
-    
-    .report-box table tr.details td {
-        padding-bottom: 20px;
-    }
-    
-    .report-box table tr.item td{
-        border-bottom: 1px solid #eee;
-    }
-    
-    .report-box table tr.item.last td {
-        border-bottom: none;
-    }
-    
-    .report-box table tr.total td:nth-child(2) {
-        border-top: 2px solid #eee;
-        font-weight: bold;
-    }
-    
-    @media only screen and (max-width: 600px) {
-        .report-box table tr.top table td {
-            width: 100%;
-            display: block;
-            text-align: center;
-        }
-        
-        .report-box table tr.information table td {
-            width: 100%;
-            display: block;
-            text-align: center;
-        }
-    }
-    
-    /** RTL **/
-    .rtl {
-        direction: rtl;
-        font-family: Tahoma, 'Helvetica Neue', 'Helvetica', Helvetica, Arial, sans-serif;
-    }
-    
-    .rtl table {
-        text-align: right;
-    }
-    
-    .rtl table tr td:nth-child(2) {
-        text-align: left;
-    }
-
-    .button {
-        background-color: white;
-        border: 2px solid #555555;
-        border-radius: 4px;
-        color: black;
-        padding: 15px 32px;
-        text-align: center;
-        text-decoration: none;
-        display: inline-block;
-        font-size: 14px;
-        transition-duration: 0.4s;
-    }
-
-    .button:hover {
-        background-color: #555555;
-        color: white;
-    }
-    </style>
-    <script src="/static/scripts.js"></script>
-    <script src="/static/jquery.min.js"></script>
-</head>
-
-<body>
-    <button class="button" onclick="window.print()">Export</button>
-    <div class="report-box">
-        <table cellpadding="0" cellspacing="0">
-            <tr class="top">
-                <td colspan="2">
-                    <table>
-                        <tr>
-                            <td class="title">
-                                <img src="https://news.region-a3.com/wp-content/uploads/sites/2/2018/04/Tat_Logo_14Cmyk.jpg" style="width:100%; max-width:300px;">
-                            </td>
-                            
-                            <td>
-                                <p id='timestamp'>TIMESTAMP</p>
-                                <p id='trainingsset'>Set: TRAININGSSET</p>
-                                <p id='quote'>0%</p>
-                            </td>
-                        </tr>
-                    </table>
-                </td>
-            </tr>
-            
-            <tr class="heading">
-                <td>
-                    Zusammenfassung
-                </td>
-                
-                <td>
-                    Anzahl
-                </td>
-            </tr>
-			
-            <tr class="item">
-                <td>
-                    Richtige Wörter
-                </td>
-                
-                <td>
-                    <p id='amt_correct'>0</p>
-                </td>
-            </tr>
-            
-			<tr class="item">
-                <td>
-                    Fast richtige Wörter
-                </td>
-                
-                <td>
-                    <p id='amt_alm_correct'>0</p>
-                </td>
-            </tr>
-			
-			<tr class="item">
-                <td>
-                    Falsche Wörter
-                </td>
-                
-                <td>
-                    <p id='amt_wrong'>0</p>
-                </td>
-            </tr>
-
-            <tr class="heading">
-                <td>
-                    Richtige Wörter
-                </td>
-				<td>
-                    
-                </td>
-            </tr>
-            
-            <tr class="item">
-                <td>
-                    <p id='correct_words'>Kein Eintrag</p>
-                </td>
-            </tr>
-            <tr class="heading">
-                <td>
-                    Fast richtige Wörter
-                </td>
-				<td>
-                    
-                </td>
-            </tr>
-            
-            <tr class="item">
-                <td>
-                    <p id='alm_correct_words'>Kein Eintrag</p>
-                </td>
-            </tr>
-            <tr class="heading">
-                <td>
-                    Falsche Wörter
-                </td>
-				<td>
-                    
-                </td>
-            </tr>
-            
-            <tr class="item">
-                <td>
-                    <p id='wrong_words'>Kein Eintrag</p>
-                </td>
-            </tr>
-        </table>
-    </div>
-</body>
-</html>
